LayoutTests/platform/ios/imported/w3c/web-platform-tests/html/browsers/the-window-object/apis-for-creating-and-navigating-browsing-contexts-by-name/open-features-non-integer-width-expected.txt


PASS top=0,left=0,height=401,: absence of feature "width" should be treated same as "width=0"
PASS top=0,left=0: absence of feature "width" should be treated same as "width=0"
PASS features "width=/404" should NOT set "width=404"
PASS features "width=_404" should NOT set "width=404"
PASS features "width=L404" should NOT set "width=404"
FAIL features "width=405.5" should set "width=405" assert_equals: "top=0,left=0,height=401,width=405.5 value after first non-digit will be ignored" expected 405 but got 980
FAIL features "width=405.32" should set "width=405" assert_equals: "top=0,left=0,height=401,width=405.32 value after first non-digit will be ignored" expected 405 but got 980
FAIL features "width=405LLl" should set "width=405" assert_equals: "top=0,left=0,height=401,width=405LLl value after first non-digit will be ignored" expected 405 but got 980
FAIL features "width=405^4" should set "width=405" assert_equals: "top=0,left=0,height=401,width=405^4 value after first non-digit will be ignored" expected 405 but got 980
FAIL features "width=405*3" should set "width=405" assert_equals: "top=0,left=0,height=401,width=405*3 value after first non-digit will be ignored" expected 405 but got 980
FAIL features "width=405/5" should set "width=405" assert_equals: "top=0,left=0,height=401,width=405/5 value after first non-digit will be ignored" expected 405 but got 980
FAIL features "width=405  " should set "width=405" assert_equals: "top=0,left=0,height=401,width=405   value after first non-digit will be ignored" expected 405 but got 980
FAIL features "width=405e1" should set "width=405" assert_equals: "top=0,left=0,height=401,width=405e1 value after first non-digit will be ignored" expected 405 but got 980
FAIL features "width=405e-1" should set "width=405" assert_equals: "top=0,left=0,height=401,width=405e-1 value after first non-digit will be ignored" expected 405 but got 980
